engine_rocks: Move DecodeProperties into engine_traits
