+++ components/engine_rocks/src/compact_listener.rs
@@ -2,7 +2,7 @@
 
 use std::cmp;
 
-use crate::properties::RangeProperties;
+use crate::properties::{RangeProperties, UserCollectedPropertiesDecoder};
 use engine::rocks::{CompactionJobInfo, EventListener};
 use tikv_util::collections::hash_set_with_capacity;
 
@@ -82,7 +82,8 @@ impl EventListener for CompactionListener {
         let mut output_props = Vec::with_capacity(info.output_file_count());
         let iter = info.table_properties().into_iter();
         for (file, properties) in iter {
-            if let Ok(prop) = RangeProperties::decode(properties.user_collected_properties()) {
+            let ucp = UserCollectedPropertiesDecoder(properties.user_collected_properties());
+            if let Ok(prop) = RangeProperties::decode(&ucp) {
                 if input_files.contains(file) {
                     input_props.push(prop);
                 } else if output_files.contains(file) {
