+++ components/engine_rocks/src/properties.rs
@@ -1,9 +1,10 @@
 // Copyright 2017 TiKV Project Authors. Licensed under Apache-2.0.
 
-use std::collections::{BTreeMap, HashMap};
+use std::collections::{HashMap};
 use std::io::Read;
 use std::ops::{Deref, DerefMut};
 use std::u64;
+use engine_traits::{IndexHandles, DecodeProperties};
 
 use engine::rocks::{
     DBEntryType, TablePropertiesCollector, TablePropertiesCollectorFactory, TitanBlobIndex,
@@ -26,64 +27,6 @@ fn get_entry_size(value: &[u8], entry_type: DBEntryType) -> std::result::Result<
             Err(_) => Err(()),
         },
         _ => Err(()),
-    }
-}
-
-#[derive(Clone, Debug, Default)]
-pub struct IndexHandle {
-    pub size: u64,   // The size of the stored block
-    pub offset: u64, // The offset of the block in the file
-}
-
-#[derive(Debug, Default)]
-pub struct IndexHandles(BTreeMap<Vec<u8>, IndexHandle>);
-
-impl Deref for IndexHandles {
-    type Target = BTreeMap<Vec<u8>, IndexHandle>;
-    fn deref(&self) -> &BTreeMap<Vec<u8>, IndexHandle> {
-        &self.0
-    }
-}
-
-impl DerefMut for IndexHandles {
-    fn deref_mut(&mut self) -> &mut BTreeMap<Vec<u8>, IndexHandle> {
-        &mut self.0
-    }
-}
-
-impl IndexHandles {
-    pub fn new() -> IndexHandles {
-        IndexHandles(BTreeMap::new())
-    }
-
-    pub fn add(&mut self, key: Vec<u8>, index_handle: IndexHandle) {
-        self.0.insert(key, index_handle);
-    }
-
-    // Format: | klen | k | v.size | v.offset |
-    fn encode(&self) -> Vec<u8> {
-        let mut buf = Vec::with_capacity(1024);
-        for (k, v) in &self.0 {
-            buf.encode_u64(k.len() as u64).unwrap();
-            buf.extend(k);
-            buf.encode_u64(v.size).unwrap();
-            buf.encode_u64(v.offset).unwrap();
-        }
-        buf
-    }
-
-    fn decode(mut buf: &[u8]) -> Result<IndexHandles> {
-        let mut res = BTreeMap::new();
-        while !buf.is_empty() {
-            let klen = number::decode_u64(&mut buf)?;
-            let mut k = vec![0; klen as usize];
-            buf.read_exact(&mut k)?;
-            let mut v = IndexHandle::default();
-            v.size = number::decode_u64(&mut buf)?;
-            v.offset = number::decode_u64(&mut buf)?;
-            res.insert(k, v);
-        }
-        Ok(IndexHandles(res))
     }
 }
 
@@ -145,20 +88,6 @@ impl UserProperties {
     }
 }
 
-pub trait DecodeProperties {
-    fn decode(&self, k: &str) -> Result<&[u8]>;
-
-    fn decode_u64(&self, k: &str) -> Result<u64> {
-        let mut buf = self.decode(k)?;
-        number::decode_u64(&mut buf)
-    }
-
-    fn decode_handles(&self, k: &str) -> Result<IndexHandles> {
-        let buf = self.decode(k)?;
-        IndexHandles::decode(buf)
-    }
-}
-
 impl DecodeProperties for UserProperties {
     fn decode(&self, k: &str) -> Result<&[u8]> {
         match self.0.get(k.as_bytes()) {
@@ -168,9 +97,13 @@ impl DecodeProperties for UserProperties {
     }
 }
 
-impl DecodeProperties for UserCollectedProperties {
+// FIXME: This is a temporary hack to implement a foreign trait on a foreign
+// type until the engine abstraction situation is straightened out.
+pub struct UserCollectedPropertiesDecoder<'a>(pub &'a UserCollectedProperties);
+
+impl<'a> DecodeProperties for UserCollectedPropertiesDecoder<'a> {
     fn decode(&self, k: &str) -> Result<&[u8]> {
-        match self.get(k.as_bytes()) {
+        match self.0.get(k.as_bytes()) {
             Some(v) => Ok(v),
             None => Err(Error::KeyNotFound),
         }
@@ -353,7 +286,7 @@ impl RangeProperties {
 impl From<SizeProperties> for RangeProperties {
     fn from(p: SizeProperties) -> RangeProperties {
         let mut res = RangeProperties::default();
-        for (key, size_handle) in p.index_handles.0 {
+        for (key, size_handle) in p.index_handles.into_map() {
             let mut range = RangeOffsets::default();
             range.size = size_handle.offset;
             res.offsets.push((key, range));
