+++ components/engine_traits/src/lib.rs
@@ -227,6 +227,9 @@ mod table_properties;
 pub use crate::table_properties::*;
 mod write_batch;
 pub use crate::write_batch::*;
+mod properties;
+pub use crate::properties::*;
+
 
 // These modules contain more general traits, some of which may be implemented
 // by multiple types.
